renamed package and other code improvements
